+++ tests/PyroTests/test_util.py
@@ -205,18 +205,19 @@ class TestUtils(unittest.TestCase):
         self.assertRaises(AttributeError,Pyro4.util.resolveDottedAttribute, obj,"a._p.q",True)    #private
         self.assertRaises(AttributeError,Pyro4.util.resolveDottedAttribute, obj,"a.__p.q",True)    #private
 
+    @unittest.skipUnless(sys.version_info>=(2,6,5), "unicode kwargs needs 2.6.5 or newer")
     def testUnicodeKwargs(self):
         # test the way the interpreter deals with unicode function kwargs
-        # those are supported by Python after 2.6.5, but not (all) by PyPy
-        # see https://bugs.pypy.org/issue751
+        # those are supported by Python after 2.6.5
         def function(*args, **kwargs):
             return args, kwargs
-        if sys.version_info>=(2,6,5):
-            processed_args=function(*(1,2,3), **{ unichr(65): 42 })
-            self.assertEqual( ((1,2,3), { unichr(65): 42}), processed_args)
-            if platform.python_implementation()!="PyPy":
-                processed_args=function(*(1,2,3), **{ unichr(0x20ac): 42 })
-                self.assertEqual( ((1,2,3), { unichr(0x20ac): 42}), processed_args)
+        processed_args=function(*(1,2,3), **{ unichr(65): 42 })
+        self.assertEqual( ((1,2,3), { unichr(65): 42}), processed_args)
+        processed_args=function(*(1,2,3), **{ unichr(0x20ac): 42 })
+        key = list(processed_args[1].keys())[0]
+        self.assertTrue(type(key) is unicode)
+        self.assertEqual(key, unichr(0x20ac))
+        self.assertEqual( ((1,2,3), { unichr(0x20ac): 42}), processed_args)
 
 
 if __name__ == "__main__":